requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
PyJWT==2.10.1
sqlparse==0.5.2
tzdata==2024.2
waitress==3.0.2
